fix: standardizing the quick-start for spring-boot school timetabling (#184)
# Description
The pull request updates the spring-boot quick start module for School
Timetabling. All domain and controller classes were adapted from the
Quarkus project to work with Spring Boot. The major changes are
described as follows:
- All REST annotations were updated to match Spring Boot annotations,
such as @RestController and @GetMapping.
- A new global error handler named `TimetableSolverExceptionHandler` has
been added.
- A new test configuration class named `ConstraintConfig` has been added
to allow injection of a bean class
`ConstraintVerifier<TimetableConstraintProvider, Timetable>`.
- The Maven and Gradle files were updated and tested.
- A new dependency called `springdoc-openapi-starter-webmvc-ui` has been
added to enable the documentation of endpoints.