chore: add mixed model (#823)
This PR updates the food packaging model to utilize the mixed model. The
proposed changes introduce a new planning entity, `operator`, and update
the entity, `line`, to use it. The current implementation already
defines an operator field, which consists of static input data, and it
has been converted into a basic variable.